+++ docs/src/components/dashboard/DashboardLight.tsx
@@ -21,6 +21,7 @@ import { SvgWidget } from "../widgets/SvgWidget";
 import useChange from "../../jacdac/useChange";
 import useWidgetTheme from "../widgets/useWidgetTheme";
 import useWidgetSize from "../widgets/useWidgetSize";
+import LightWidget from "../widgets/LightWidget";
 /*
 0xD6: range P=0 N=length W=1 S=0- range from pixel P, Npixels long (currently unsupported: every Wpixels skip Spixels)
 */
@@ -187,122 +188,6 @@ function LightCommand(props: { service: JDService, expanded: boolean }) {
             </IconButtonWithTooltip>
         </Grid>
     </Grid>
-}
-
-function rgbToHsl(r: number, g: number, b: number): [number, number, number] {
-    const [r$, g$, b$] = [r / 255, g / 255, b / 255];
-    let cMin = Math.min(r$, g$, b$);
-    let cMax = Math.max(r$, g$, b$);
-    let cDelta = cMax - cMin;
-    let h: number;
-    let s: number;
-    let l: number;
-    let maxAndMin = cMax + cMin;
-
-    //lum
-    l = (maxAndMin / 2) * 100
-
-    if (cDelta === 0)
-        s = h = 0;
-    else {
-        //hue
-        if (cMax === r$)
-            h = 60 * (((g$ - b$) / cDelta) % 6);
-        else if (cMax === g$)
-            h = 60 * (((b$ - r$) / cDelta) + 2);
-        else if (cMax === b$)
-            h = 60 * (((r$ - g$) / cDelta) + 4);
-
-        //sat
-        if (l > 50)
-            s = 100 * (cDelta / (2 - maxAndMin));
-        else
-            s = 100 * (cDelta / maxAndMin);
-    }
-
-    return [Math.floor(h), Math.floor(s), Math.floor(l)];
-}
-
-function setRgb(el: SVGElement, r: number, g: number, b: number) {
-    const hsl = rgbToHsl(r, g, b);
-    const [h, s, l] = hsl;
-    // at least 70% luminosity
-    const lum = Math.max(l, 60);
-    const fill = `hsl(${h}, ${s}%, ${lum}%)`;
-    el.setAttribute("fill", fill);
-}
-
-function LightWidget(props: DashboardServiceProps) {
-    const { service } = props;
-    const host = useServiceHost<LightServiceHost>(service);
-    const { background, controlBackground } = useWidgetTheme()
-    const widgetSize = useWidgetSize()
-    const [numPixels] = useChange(host.numPixels, r => r.values<[number]>());
-    const [variant] = useChange(host.variant, r => r.values<[LightVariant]>());
-    const pathRef = useRef<SVGPathElement>(undefined)
-    const pixelsRef = useRef<SVGGElement>(undefined);
-
-    const neoradius = 6;
-    const neoperimeter = numPixels * (3 * neoradius)
-    const ringradius = neoperimeter / (2 * Math.PI)
-    const margin = 2 * neoradius;
-    const width = 2 * (margin + ringradius);
-    const height = width;
-    const neocircleradius = neoradius + 1;
-    const sw = margin;
-    const wm = width - 2 * margin;
-    let d = "";
-    //if (variant === LightVariant.Ring)
-    d = `M ${margin},${height >> 1} a ${ringradius},${ringradius} 0 1,0 ${wm},0 a ${ringradius},${ringradius} 0 1,0 -${wm},0`
-
-    // paint svg via dom
-    const render = () => {
-        const colors = host.colors;
-        const pixels = pixelsRef.current.children;
-        const pn = Math.min(pixels.length, colors.length / 3);
-        let ci = 0;
-        for (let i = 0; i < pn; ++i) {
-            const pixel = pixels.item(i) as SVGCircleElement;
-            setRgb(pixel, colors[ci], colors[ci + 1], colors[ci + 2]);
-            ci += 3;
-        }
-    }
-
-    // reposition pixels along the path
-    useEffect(() => {
-        const p = pathRef.current;
-        const pixels = pixelsRef.current.children;
-        const pn = pixels.length;
-        const length = p.getTotalLength();
-        const extra = variant === LightVariant.Ring ? 0 : 1;
-        const step = length / (pn + extra);
-
-        for (let i = 0; i < pn; ++i) {
-            const pixel = pixels.item(i) as SVGCircleElement;
-            const point = p.getPointAtLength(step * (i + (extra >> 1)));
-            pixel.setAttribute("cx", "" + point.x);
-            pixel.setAttribute("cy", "" + point.y);
-        }
-
-        render();
-    }, [variant, numPixels])
-
-    // render when new colors are in
-    useEffect(() => host.subscribe(RENDER, render), [host]);
-
-    return <SvgWidget width={width} height={height} size={widgetSize}>
-        <>
-            <path ref={pathRef} d={d} fill="transparent" stroke={background} strokeWidth={sw} />
-            <g ref={pixelsRef}>
-                {Array(numPixels).fill(0).map((_, i) => <circle key={"pixel" + i}
-                    r={neocircleradius}
-                    cx={width >> 1} cy={height >> 1}
-                    stroke={controlBackground}
-                    strokeWidth={1}
-                />)}
-            </g>
-        </>
-    </SvgWidget>
 }
 
 export default function DashboardLight(props: DashboardServiceProps) {
